feat(01_vcs): split the code into files and included ifndef
