+++ Dis-Artist/01_vcs/vcs.h
@@ -1,18 +1,26 @@
+#ifndef VCH_H
+#define VCH_H
+
 #include <unistd.h>
 #include <stdio.h>
-#include <sys/socket.h>
 #include <stdlib.h>
-#include <netinet/in.h>
 #include <string.h>
 #include <fstream>
 #include <iostream>
 #include <vector>
-#include <arpa/inet.h>
 #include <filesystem>
 
-#define PORT 8050
-#define CHUNK_SIZE 1024
 using std::string;
 using std::endl;
 using std::cout;
 using std::vector;
+
+void commit(std::string message);
+void show_log();
+void checkout_commit_id(string commit_id);
+void create_branch_and_checkout(string branch_name);
+void display_branches();
+void merge(string branch_2, string branch_1);
+void display_help();
+
+#endif
